Some possible nav solutions
Hey, check out the comments in the file, but basically I fixed the nav
making it ‘sticky’ and also taking it out of the flow so the hero is
positioned the way you want. Also, ‘visibility’ will probably help
solve the logo image/text on small. Let me know if you have questions.